+++ src/schemathesis/contrib/__init__.py
@@ -1,1 +1,11 @@
-from . import unique_data
+from . import openapi, unique_data
+
+
+def install() -> None:
+    openapi.install()
+    unique_data.install()
+
+
+def uninstall() -> None:
+    openapi.uninstall()
+    unique_data.uninstall()